move Digpoint to digitization
