+++ mne/digitization/__init__.py
@@ -1,0 +1,13 @@
+from .base import DigPoint
+
+# from .base import (_dig_kind_dict, _dig_kind_ints, _dig_kind_proper,
+#                    _dig_kind_rev, _cardinal_kind_rev,)
+
+__all__ = [
+    'DigPoint',
+    # '_dig_kind_dict',
+    # '_dig_kind_ints',
+    # '_dig_kind_proper',
+    # '_dig_kind_rev',
+    # '_cardinal_kind_rev',
+]
