+++ mne/io/__init__.py
@@ -8,7 +8,7 @@
 from .open import fiff_open, show_fiff, _fiff_get_fid
 from .meas_info import (read_fiducials, write_fiducials, read_info, write_info,
                         _empty_info, _merge_info, _force_update_info, Info,
-                        anonymize_info, _stamp_to_dt, DigPoint)
+                        anonymize_info, _stamp_to_dt)
 
 from .proj import make_eeg_average_ref_proj, Projection
 from .tag import _loc_to_coil_trans, _coil_trans_to_loc, _loc_to_eeg_loc
@@ -51,3 +51,19 @@ from .fiff import Raw as RawFIF
 from .base import concatenate_raws
 from .reference import (set_eeg_reference, set_bipolar_reference,
                         add_reference_channels)
+
+__all__ = [
+    _coil_trans_to_loc, _empty_info, _fiff_get_fid, _force_update_info,
+    _loc_to_coil_trans, _loc_to_eeg_loc, _merge_info, _stamp_to_dt,
+    BaseRaw, Info, Projection, Raw, RawArray, RawFIF, add_reference_channels,
+    anonymize_info, array, base, brainvision, bti, cnt, concatenate_raws,
+    constants, ctf, edf, eeglab, egi, fiff, fiff_open, find_edf_events, kit,
+    make_eeg_average_ref_proj, nicolet, pick, read_epochs_eeglab,
+    read_epochs_fieldtrip, read_epochs_kit, read_evoked_fieldtrip,
+    read_fiducials, read_info, read_raw_artemis123, read_raw_bdf,
+    read_raw_brainvision, read_raw_bti, read_raw_cnt, read_raw_ctf,
+    read_raw_edf, read_raw_eeglab, read_raw_egi, read_raw_eximia,
+    read_raw_fieldtrip, read_raw_fif, read_raw_gdf, read_raw_kit,
+    read_raw_nicolet, set_bipolar_reference, set_eeg_reference, show_fiff,
+    write_fiducials, write_info,
+]
